lk2nd: scripts: mkbootimg: Add option for QCDT
QCDT is Qualcomm's custom DTB container format, that allows automatic
selection of a DTB based on certain board properties. Unfortunately
it makes use of certain reserved fields of the Android boot image
format that were later assigned for different use (e.g. the header
version), so it is not compatible with newer Android boot images.