platform: *: gpio: Rename conflicting gpio_set
gpio_set is defined in the lk gpio interface but some platforms
reuse the name for caf lk's internal gpio implementation. Rename the
function to gpio_set_dir as used by some other platforms to remove
the conflict.

The msm7XXX and qsdXXXX are kept as is as they implement the lk api
properly